changed masks to be 64 bits

+++ src/masks.c
@@ -1,8 +1,8 @@
 #include "masks.h"
 
-uint32_t build_mask(int start_bit, int end_bit)
+uint64_t build_mask(int start_bit, int end_bit)
 {
-    uint32_t mask = 0;
+    uint64_t mask = 0;
 
     for (int bit = start_bit; bit <= end_bit; bit++)
     {
