+++ src/masks.h
@@ -1,4 +1,4 @@
 #pragma once
 #include <stdint.h>
 
-uint32_t build_mask(int start_bit, int end_bit);
+uint64_t build_mask(int start_bit, int end_bit);
